+++ internal/safefile/root_id_windows.go
@@ -1,0 +1,20 @@
+package safefile
+
+import (
+	"fmt"
+	"golang.org/x/sys/windows"
+	"os"
+)
+
+func PersistentRootID(path string) (string, error) {
+	f, err := os.Open(path)
+	if err != nil {
+		return "", err
+	}
+	defer f.Close()
+	var info windows.ByHandleFileInformation
+	if err := windows.GetFileInformationByHandle(windows.Handle(f.Fd()), &info); err != nil {
+		return "", err
+	}
+	return fmt.Sprintf("%d:%d:%d", info.VolumeSerialNumber, info.FileIndexHigh, info.FileIndexLow), nil
+}